+++ packages/api/src/middlewares/auth.middleware.ts
@@ -4,7 +4,7 @@ import { ExpressMiddlewareInterface } from 'routing-controllers';
 import { User } from 'src/entities/user';
 
 export interface RequestWithUser extends Request {
-  userId?: string;
+  userId: string;
   user?: Partial<User>;
 }
 
@@ -15,9 +15,7 @@ export class authMiddleware implements ExpressMiddlewareInterface {
     const token = accessToken.split('Bearer ')[1];
     if (!token) return response.sendStatus(401);
     try {
-      const data = <{ userId: string }>(
-        verify(token, process.env.ACCESS_TOKEN_SECRET as string)
-      );
+      const data = <{ userId: string }>verify(token, process.env.ACCESS_TOKEN_SECRET as string);
       request.userId = data.userId;
       return next();
     } catch {